requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
PyJWT>=2.9.0
pg8000>=1.31.0
